parser.go: Fix Bird 2.0 BGP protocol parsing
In Bird 2.0, BGP protocol lines have `---` as the table name, as the actual
table names are defined in the respective channels. This commit fixes the
regex to take this into account.

Parsing of the key-value section is also changed so that we don't mistake
the states of individual channels as the protocol state. In the future, we
should distinguish between different channels instead of parsing the key-value
section blindly.

Example:

```
xxx1_4     BGP        ---        up     2018-08-21 08:44:30  Established
  Description:    iBGP xxx1 v4
  BGP state:          Established
  <snip>
  Channel ipv4
    State:          UP
    Table:          master4
  <snip>
```

- Expected `state`: `up`
- Parsed `state`: `UP` (from the ipv4 channel)

- Expected `state_changed`: `2018-08-21 08:44:30`
- Parsed `state_changed`: Not parsed